Interpreter-only removal of preemption blockers
This change breaks the compiler. If you apply this patch you must
disable the compiler and use only the interpreter.

The goal of this change is to bolster the VM's ability to respond to
events quickly, and ideally to make 60fps achievable on the devices that
are capable of running the VM.

To accomplish this, we reduce the amount of time that individual threads
are allowed to run before being preempted. However, simply reducing that
time is not enough: There are a few mechanisms in place that block
preemptions from happening, so threads can run arbitrarily long
depending on whether they are initializing classes, running
compiled/interpreted/native code, throwing exceptions, etc.

Remove `executeFrame`. This is a function that runs a specified amount
of code through the interpreter and does not allow preemptions while
that code is executing. All callers are updated to instead push a frame
to the call stack and to no longer expect return values to be available
synchronously.

Change how Java Exceptions are created and thrown. Exceptions were being
created using the `createException` function on the `Context` class.
This function is now removed. It used to block preemptions, call the
constructor of a Java Exception subclass directly and return the result.
Callers that want to throw exceptions now push the Exception constructor
to the frame stack and continue execution.

Change how yields/preemptions work. We were explicitly checking at
various points whether we should preempt and, if so, setting the `U`
property to indicate that the current thread should unwind. Instead, the
interpreter now checks on every step whether it should preempt and, if
so, just leaves the interpret loop. All thread contexts hold the
information necessary to resume from a preemption without needing to
explicitly save/load anything.

Change how class initialization works. Class initialization was one of
the places where we were frequently calling into interpreted Java code
without the possibility of preemption. Call sites that need to
initialize classes now push class initialization onto the call stack and
don't expect return values to be available synchronously.